Crate http_async

Simple Server, who speaks the Hyper Text Transfer Protocol, with async-std.

Usage

Include this:

  [dependencies]
  http-async = "0.1"

in you project Cargo.toml-file.

Licence

MIT

Modules

header
Header of Requests and Responses.

method
Protocol Methods.

path
Path and Queries to Resources.

request
Requests to Server.

response
Responses to Client.

status
Status Codes

version
Protocol Versions.

Macros

content
Just send this content successfully.

Structs

Configuration
Configuration of the HTTP Server.

Content
Content of a Hyper Text Transfer Protocol Response.

KeyValuePair
Simple Key-Value-Pair. E.g. for header-fields, Queries, etc.

Functions

Content
Constructor for Content.

server
Creates a Future to start a Hyper Text Transfer Protocol Server.