Struct http_async::request::Request

pub struct Request {
    pub method: Method,
    pub path: String,
    pub querySeperator: char,
    pub query: Vec<KeyValuePair>,
    pub version: Version,
    pub header: Vec<KeyValuePair>,
    pub content: Vec<u8>,
}

Hyper Text Transfer Protocol Request.

Fields
method: Method

Method to use to get the desired Resource.

path: String

Path to the desired Resource.

querySeperator: char

When parsing a Request, how should I differentiate the Key Value Pairs of the Query?

query: Vec<KeyValuePair>

List of Query Key Value Pairs.

version: Version

Protocol Version.

header: Vec<KeyValuePair>

List of Header Key Value Pairs.

content: Vec<u8>

Content of Request.

Methods

impl Request

pub fn ifChar<Inner>(
character: char,
stream: &mut TcpStream,
inner: Inner
) -> Option<Inner>

Read char from stream and compare it with character.

Arguments

character – char to compare with,

stream – Transmission Control Protocol Stream,

inner – inner value to return on success.

pub fn readChar(stream: &mut TcpStream) -> Option<char>

Try to read a single char from Transmission Control Protocol stream.

Arguments

stream – Transmission Control Protocol Stream.

pub async fn parse<'_>(
__arg0: Self,
__arg1: &'_ mut TcpStream
) -> Result<Self, String>

Parse TcpStream as Hyper Text Transfer Protocol Request.

Arguments

stream – Transmission Control Protocol Stream.