Struct http_async::Configuration

pub struct Configuration {
    pub address: SocketAddr,
    pub maxContent: usize,
    pub maxHeaderEntries: usize,
    pub maxHeaderLength: usize,
    pub maxPathLength: usize,
    pub maxQueryLength: usize,
}

Configuration of the HTTP Server.

Fields
address: SocketAddr

Address, on which the Server should listen to.

maxContent: usize

Maximum Bytes of Request Content. Prevents DOS by Allocation a large Buffer (Header ›Content-Length‹ could contain any decimal value) without ever filling it.

maxHeaderEntries: usize

Maximum Numbers of Headers. Prevents Slow Lorris Attacks: Client might slowly send Header by Header for ever, but because neither the Connection times out nor the Request every ends, the Server keeps reading the Stream.

maxHeaderLength: usize

Maximum Length of a Header.

maxPathLength: usize

Maximum Length of Path. Prevents Slow Lorris Attacks.

maxQueryLength: usize

Maximum Length of Query String. Prevents Slow Lorris Attacks.